pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ncut_pytorch"
version = "2.0.0"
authors = [
  { name="Huzheng Yang", email="huze.yann@gmail.com" },
]
description = "Normalized Cut and Nyström Approximation"
readme = "README.md"
requires-python = ">=3"
license = "MIT"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Mathematics",
]
dependencies = [
    "torch~=2.0",
    "fpsample>=0.2.0",
    "pytorch-lightning~=2.0",
    "pillow",
    "numpy",
    "tqdm"
]

[project.optional-dependencies]
vision = [
    "torchvision>=0.15.0",
    "opencv-python",
    "matplotlib",
]

[project.urls]
Documentation = "https://ncut-pytorch.readthedocs.io/"
Github = "https://github.com/huzeyann/ncut_pytorch/"
PyPI = "https://pypi.org/project/ncut_pytorch/"

[tool.setuptools]
packages = ["ncut_pytorch", "ncut_pytorch.ncuts", "ncut_pytorch.color", "ncut_pytorch.utils", "ncut_pytorch.dino"]

[tool.setuptools.package-data]
"ncut_pytorch" = ["*.txt", "*.md"]